FISA_Link_website/requirements.txt
mysqlclient
django-cors-headers
mysql-connector-python
